*NOT HEAVILY TESTED*
Added feature to export API keys properly.  There is now a new Python script that handles serving the *.html boilerplate files.  Depending on the domain it is on, it will serve the proper API key.  The API keys are stored in the boilerplateProxy.py file.  Also, when you hit "export code" it will use the API keys listed in interactive_logic.js to export to savedbythegoog.appspot.com with the correct API key.  Lastly, if you hit "view source", it will display the key as &key=<<INSERT KEY>>.

Added a way to detect clicks in the code editor, in editor.js.

Added shadows, draggable, resizable to the Pick & Edit divs.  Also made a handler so that the div you click on "comes to front".  Took out +/- buttons

Fixed "code menu" -- hovering over/out does the predictable hiding of the menu.

Added localsearch control samples.